+++ app/src/main/java/com/busanit501/androidlabtest501/miniProject/lsy1205_mini/Ch11MainRecyclerView1Activity.kt
@@ -10,7 +10,7 @@ import com.busanit501.androidlabtest501.R
 import com.busanit501.androidlabtest501.ch11_jetpack.recyclerview.test1simple.Ch11MyAdapterSample_msy0402
 import com.busanit501.androidlabtest501.databinding.ActivityCh11MainRecyclerView1Binding
 
-class Ch11MainRecyclerView1Activity_msy0402 : AppCompatActivity() {
+class Ch11MainRecyclerView1Activity : AppCompatActivity() {
     lateinit var binding: ActivityCh11MainRecyclerView1Binding
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -32,7 +32,7 @@ class Ch11MainRecyclerView1Activity_msy0402 : AppCompatActivity() {
         }
 
         // Set layout manager
-        val layoutManager = LinearLayoutManager(this@Ch11MainRecyclerView1Activity_msy0402)
+        val layoutManager = LinearLayoutManager(this@Ch11MainRecyclerView1Activity)
         layoutManager.orientation = LinearLayoutManager.VERTICAL
         binding.ch11RecyclerSample1.layoutManager = layoutManager
 
@@ -46,7 +46,7 @@ class Ch11MainRecyclerView1Activity_msy0402 : AppCompatActivity() {
         }
 
         // Set layout manager and adapter for the second RecyclerView
-        val layoutManager2 = LinearLayoutManager(this@Ch11MainRecyclerView1Activity_msy0402)
+        val layoutManager2 = LinearLayoutManager(this@Ch11MainRecyclerView1Activity)
         layoutManager2.orientation = LinearLayoutManager.VERTICAL
         binding.ch11RecyclerSample2.layoutManager = layoutManager2
         binding.ch11RecyclerSample2.adapter = Ch11MyAdapterSample_msy0402(datas2)
@@ -58,7 +58,7 @@ class Ch11MainRecyclerView1Activity_msy0402 : AppCompatActivity() {
         }
 
         // Set layout manager and adapter for the third RecyclerView
-        val layoutManager3 = LinearLayoutManager(this@Ch11MainRecyclerView1Activity_msy0402)
+        val layoutManager3 = LinearLayoutManager(this@Ch11MainRecyclerView1Activity)
         layoutManager3.orientation = LinearLayoutManager.VERTICAL
         binding.ch11RecyclerSample3.layoutManager = layoutManager3
         binding.ch11RecyclerSample3.adapter = Ch11MyAdapterSample_msy0402(datas3)
