+++ app/src/main/java/com/busanit501/androidlabtest501/miniProject/test0709/msy0402_mini/Ch11MyViewHolder2_msy0402.kt
@@ -1,0 +1,7 @@
+package com.busanit501.androidlabtest501.ch11_jetpack.recyclerview.test1simple
+
+import androidx.recyclerview.widget.RecyclerView
+import com.busanit501.androidlabtest501.databinding.Ch11ItemSimpleRecycler2Binding
+
+class Ch11MyViewHolder2_msy0402(val binding: Ch11ItemSimpleRecycler2Binding): RecyclerView.ViewHolder(binding.root) {
+}
