+++ app/src/main/java/com/busanit501/androidlabtest501/miniProject/test0709/msy0402_mini/Ch11MyViewHolder_msy0402.kt
@@ -1,0 +1,14 @@
+package com.busanit501.androidlabtest501.ch11_jetpack.recyclerview.test1simple
+
+import androidx.recyclerview.widget.RecyclerView
+import com.busanit501.androidlabtest501.databinding.Ch11ItemSimpleRecyclerBinding
+
+// 순서 2
+// 목록의 요소 뷰를 생성.
+
+// 순서3
+// ch11_item_simple_recycler , 파일 생성, layout 폴더
+//
+class Ch11MyViewHolder_msy0402(val binding:Ch11ItemSimpleRecyclerBinding ):RecyclerView.ViewHolder(binding.root) {
+
+}
